feat: minute flushing plus oversize warning + refactor files
